+++ shared/storages/task/task.go
@@ -9,7 +9,7 @@ type Task struct {
 	Id          int64      `json:"id"`
 	Title       string     `sql:"default: not null", json:"title"`
 	Description *string    `sql:"default: null", json:"description"`
-	Priority    *int64     `sql:"default: null", json:"priority"`
+	Priority    *int       `sql:"default: null", json:"priority"`
 	CreatedAt   time.Time  `json:"createdAt"`
 	UpdatedAt   time.Time  `json:"updatedAt"`
 	CompletedAt *time.Time `json:"completedAt"`
